requirements.txt
django
gunicorn
django-heroku
requests
pillow
django-ckeditor